feat: Update playwright and refactor to unified server
- Updated the `playwright` dependency to the latest version (`1.45.0`).
- Refactored the addon to use a unified server architecture, removing the obsolete "bridge" microservice.
- Moved all stream extraction logic from `server.js` to `lib/handlers.js`.
- Implemented a `getBrowser` function to safely pass the Playwright browser instance to the handlers.
- Fixed a bug in the series URL resolution logic to prioritize exact matches, improving scraping accuracy.
- Cleaned up `server.js` by removing the now-redundant `/extract-streams` endpoint and its associated code.
