+++ lib/handlers.js
@@ -7,7 +7,88 @@ const {
     normalizeId
 } = require("./scraping");
 
-function defineHandlers(builder) {
+const NodeCache = require('node-cache');
+const streamCache = new NodeCache({ stdTTL: 3600, checkperiod: 600 });
+
+const PROVIDERS = {
+    'yourupload.com': async (page) => {
+        await page.waitForSelector('video');
+        return page.evaluate(() => document.querySelector('video')?.src || document.querySelector('video source')?.src);
+    },
+    'mp4upload.com': async (page) => {
+        await page.waitForSelector('video', { state: 'visible', timeout: 20000 });
+        return page.evaluate(() => {
+            const scripts = Array.from(document.querySelectorAll('script'));
+            for (const script of scripts) {
+                const match = script.textContent.match(/https?:\/\/[^"']+\.(mp4|m3u8)[^"']*/);
+                if (match) return match[0];
+            }
+            const video = document.querySelector('video');
+            return video?.src || video?.querySelector('source')?.src;
+        });
+    },
+    'vidsrc.to': async (page) => {
+        await page.waitForSelector('iframe');
+        const iframeSrc = await page.$eval('iframe', el => el.src);
+        if (iframeSrc.includes('m3u8')) return iframeSrc;
+        return page.evaluate(() => {
+            const scripts = document.querySelectorAll('script');
+            for (const script of scripts) {
+                const match = script.textContent.match(/https?:\/\/[^"']+\.m3u8[^"']*/);
+                if (match) return match[0];
+            }
+            return null;
+        });
+    }
+};
+
+async function extractVideoUrl(browser, url, referer = null) {
+    const cacheKey = `video_url:${url}`;
+    const cached = streamCache.get(cacheKey);
+    if (cached) {
+        console.log(`Cache hit for final URL: ${url}`);
+        return cached;
+    }
+
+    const page = await browser.newPage();
+    try {
+        await page.setExtraHTTPHeaders({
+            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36',
+            'Referer': referer || new URL(url).origin,
+        });
+        await page.goto(url, { waitUntil: 'networkidle', timeout: 30000 });
+
+        const detectedProvider = Object.keys(PROVIDERS).find(p => url.includes(p));
+        const extractor = detectedProvider ? PROVIDERS[detectedProvider] : null;
+
+        let videoUrl;
+        if (extractor) {
+            videoUrl = await extractor(page);
+        } else {
+            videoUrl = await page.evaluate(() => {
+                const video = document.querySelector('video');
+                if (video?.src) return video.src;
+                const source = document.querySelector('video source');
+                if (source?.src) return source.src;
+                const scripts = Array.from(document.querySelectorAll('script'));
+                for (const script of scripts) {
+                    const match = script.textContent.match(/https?:\/\/[^\s"']+\.(?:m3u8|mp4)[^\s"']*/);
+                    if (match) return match[0];
+                }
+                return null;
+            });
+        }
+
+        if (videoUrl) {
+            streamCache.set(cacheKey, videoUrl);
+        }
+        return videoUrl;
+    } finally {
+        await page.close();
+    }
+}
+
+function defineHandlers(builder, getBrowser) {
     builder.defineCatalogHandler(async ({type, id, extra}) => {
         console.log("request for catalog: " + type + " " + id);
 
@@ -174,8 +255,6 @@ function defineHandlers(builder) {
         }
     });
 
-    const BRIDGE_URL = process.env.BRIDGE_URL || 'http://localhost:3001';
-
     async function getLatanimeEpisodeUrl(imdbId, season, episode) {
         try {
             const cinemetaUrl = `https://v3-cinemeta.strem.io/meta/series/${imdbId}.json`;
@@ -196,22 +275,25 @@ function defineHandlers(builder) {
                 } else {
                     const seasonRegex = new RegExp(`(temporada|season|s)[\\s\\-]*${season}\\b`, 'i');
                     const exactMatchRegex = new RegExp(`^${seriesName}$`, 'i');
-                    let bestMatch = null;
+
+                    let perfectMatch = null;
+                    let seasonMatch = null;
 
                     searchResults.each((i, el) => {
                         const link = $(el);
                         const title = link.find('h3').text().trim();
                         const href = link.attr('href');
 
-                        if (season > 1 && (title.match(seasonRegex) || href.match(seasonRegex))) {
-                            bestMatch = href;
-                            return false;
+                        if (title.match(exactMatchRegex)) {
+                            perfectMatch = href;
+                            return false; // Stop searching once a perfect match is found
                         }
-                        if (season === 1 && title.match(exactMatchRegex)) {
-                            bestMatch = href;
+                        if (season > 1 && (title.match(seasonRegex) || href.match(seasonRegex))) {
+                           if (!seasonMatch) seasonMatch = href;
                         }
                     });
-                    seriesUrl = bestMatch || searchResults.first().attr('href');
+
+                    seriesUrl = perfectMatch || seasonMatch || searchResults.first().attr('href');
                 }
             }
 
@@ -278,18 +360,101 @@ function defineHandlers(builder) {
             targetUrl = `${LATANIME_URL}/ver/${episodeId}`;
         }
 
-        try {
-            const response = await axios.post(`${BRIDGE_URL}/extract-streams`, { url: targetUrl }, { timeout: 120000 });
-            if (response.data && response.data.streams) {
-                console.log(`Received ${response.data.streams.length} streams from bridge`);
-                return Promise.resolve({ streams: response.data.streams });
-            } else {
-                console.log('Bridge returned no streams');
-                return Promise.resolve({ streams: [] });
-            }
-        } catch (error) {
-            console.error(`Error communicating with bridge for ${targetUrl}:`, error.message);
+        const browser = getBrowser();
+        if (!browser) {
+            console.error('Browser is not ready');
             return Promise.resolve({ streams: [] });
+        }
+
+        console.log(`Scraping latanime page: ${targetUrl}`);
+        const page = await browser.newPage();
+        try {
+            await page.goto(targetUrl, { waitUntil: 'domcontentloaded' });
+
+            const providers = await page.evaluate(() => {
+                const results = [];
+                const baseKey = document.querySelector('div.player')?.getAttribute('data-key');
+                if (!baseKey) return [];
+                const basePlayerUrl = atob(baseKey);
+                document.querySelectorAll('a.play-video').forEach(el => {
+                    const providerName = el.textContent.trim();
+                    const encodedPart = el.getAttribute('data-player');
+                    if (encodedPart) {
+                        const intermediateUrl = providerName.toLowerCase() === 'yourupload' ? atob(encodedPart) : basePlayerUrl + encodedPart;
+                        results.push({ url: intermediateUrl, title: providerName });
+                    }
+                });
+                return results;
+            });
+            console.log(`Found ${providers.length} potential providers.`);
+
+            const finalEmbedUrls = await Promise.all(providers.map(async (provider) => {
+                const providerPage = await browser.newPage();
+                try {
+                    await providerPage.goto(provider.url, { waitUntil: 'domcontentloaded' });
+                    const finalUrl = await providerPage.evaluate(() => {
+                        const redirMatch = document.body.innerHTML.match(/var redir = atob\("([^"]+)"\);/);
+                        return redirMatch ? atob(redirMatch[1]) : null;
+                    });
+                    return { ...provider, finalUrl };
+                } catch (e) {
+                    console.error(`Failed to get final embed from ${provider.url}: ${e.message}`);
+                    return null;
+                } finally {
+                    await providerPage.close();
+                }
+            }));
+            const validEmbeds = finalEmbedUrls.filter(p => p && p.finalUrl && !p.finalUrl.includes('listeamed.net'));
+            console.log(`Found ${validEmbeds.length} valid final embed URLs (after filtering).`);
+
+            const streamPromises = validEmbeds.map(async (provider) => {
+                try {
+                    const videoUrl = await extractVideoUrl(browser, provider.finalUrl, provider.url);
+                    if (videoUrl) {
+                        console.log(`✅ Extracted: ${provider.title} -> ${videoUrl.substring(0, 60)}...`);
+                        return {
+                            name: 'Latanime',
+                            url: videoUrl,
+                            title: provider.title,
+                            behaviorHints: {
+                                proxyHeaders: {
+                                    'request': {
+                                        'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36',
+                                        'Referer': new URL(provider.finalUrl).origin
+                                    }
+                                }
+                            }
+                        };
+                    }
+                } catch (error) {
+                    console.log(`❌ Extraction failed for ${provider.finalUrl}: ${error.message}`);
+                }
+                return null;
+            });
+            const resolvedStreams = (await Promise.all(streamPromises)).filter(Boolean);
+
+            const downloadLinks = await page.evaluate(() => {
+                const links = [];
+                const selectors = [
+                    'a[href*="pixeldrain.com"]', 'a[href*="mediafire.com"]', 'a[href*="mega.nz"]',
+                    'a[href*="gofile.io"]', 'a[href*="drive.google.com"]', 'a[href*="1fichier.com"]',
+                    'a[download]'
+                ];
+                document.querySelectorAll(selectors.join(',')).forEach(el => {
+                    if (el.href) links.push({ url: el.href, title: `📥 ${el.textContent.trim() || 'Download'}` });
+                });
+                return links;
+            });
+
+            const allStreams = [...resolvedStreams, ...downloadLinks];
+            console.log(`Total streams found: ${allStreams.length}`);
+            return Promise.resolve({ streams: allStreams });
+
+        } catch (error) {
+            console.error(`Scraping error on ${targetUrl}: ${error.message}`);
+            return Promise.resolve({ streams: [] });
+        } finally {
+            await page.close();
         }
     });
 }
