+++ package.json
@@ -12,7 +12,7 @@
     "cors": "^2.8.5",
     "express": "^4.19.2",
     "node-cache": "^5.1.2",
-    "playwright": "1.44.0",
+    "playwright": "1.45.0",
     "scrapingbee": "^1.7.6",
     "stremio-addon-sdk": "^1.6.10"
   }
